Move containsMissingBlock util to separate helper util
