+++ apps/full-site-editing/full-site-editing-plugin/starter-page-templates/page-template-modal/index.js
@@ -22,36 +22,10 @@ import { trackDismiss, trackSelection, trackView } from './utils/tracking';
 import replacePlaceholders from './utils/replace-placeholders';
 import ensureAssets from './utils/ensure-assets';
 import mapBlocksRecursively from './utils/map-blocks-recursively';
+import containsMissingBlock from './utils/contains-missing-block';
 /* eslint-enable import/no-extraneous-dependencies */
 
 const DEFAULT_HOMEPAGE_TEMPLATE = 'maywood';
-
-/**
- * Determines whether the provided collection of Blocks contains any "missing"
- * blocks as determined by the presence of the `core/missing` block type.
- *
- * @param {Array} blocks the collection of block objects to check for "missing" block .
- * @returns {boolean} whether the collection blocks contains any missing blocks.
- */
-function containsMissingBlock( blocks ) {
-	// Once parsed, missing Blocks have a name prop of `core/missing`.
-	// see: https://github.com/WordPress/gutenberg/tree/742dbf2ef0e37481a3c14c29f3688aa0cd3cf887/packages/block-library/src/missing
-	const MISSING_BLOCK_NAME = 'core/missing';
-
-	return blocks.find( block => {
-		// If we found a missing block the bale out immediately
-		if ( block.name === MISSING_BLOCK_NAME ) {
-			return true;
-		}
-
-		// If there are innerblocks then recurse down into them...
-		if ( block.innerBlocks && block.innerBlocks.length ) {
-			return containsMissingBlock( block.innerBlocks );
-		}
-
-		return false;
-	} );
-}
 
 class PageTemplateModal extends Component {
 	state = {
